requirements.txt
asttokens==2.4.1
bayesian-optimization==1.4.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.2.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
fonttools==4.38.0
fsspec==2024.3.1
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
jupyter-client==8.6.0
jupyter-core==5.5.0
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.6
mpmath==1.3.0
mysql-connector-python==8.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.24.1
packaging==23.0
pandas==2.1.3
parso==0.8.3
peewee==3.16.3
peewee-migrate==1.12.2
Pillow==9.4.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pyaml==23.7.0
pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.28.1
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.10.1
six==1.16.0
sklearn==0.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.1.0
torch==2.2.1
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.0
typing-extensions==4.10.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.12
